Adding correct dependencies
