--- ingestion/src/main/scala/uk/gov/ons/bi/ingest/records/parsers/OffsetProvider.scala
@@ -1,8 +1,0 @@
-package uk.gov.ons.bi.ingest.records.parsers
-
-trait OffsetProvider[T] extends RecordParser[T] {
-
-  def parser: OffsetParser
-
-  override type SourceType = Map[String, String]
-}
