+++ models/src/main/scala/uk/gov/ons/bi/ingest/parsers/OffsetParser.scala
@@ -1,0 +1,8 @@
+package uk.gov.ons.bi.ingest.parsers
+
+trait OffsetParser[T] extends RecordParser[T] {
+
+  def parser: OffsetFormat
+
+  override type SourceType = Map[String, String]
+}
